requirements.txt
accelerate==0.21.0
bitsandbytes==0.40.2
chromadb
git+https://github.com/openai/whisper.git
google-generativeai
gradio
huggingface-hub
langchain
langchain_community
langchain-google-genai
paddleocr
paddlepaddle-gpu
peft==0.4.0
sacremoses
speechrecognition
streamlit
transformers==4.31.0
trl==0.4.7
streamlit-mic-recorder
